refactor(niri): separate running a body from reading one
The frame performs each statement exactly once, over values the engine itself
built. The lattice projection is re-entrant and may run over the same node any
number of times. Both lived in one `impl`, and the distinction survived only as
prose — which is how a mutating call came to be folded in the projection, where
its write would land twice or not at all.

`frame` now holds the executor and `Flow`, and reaches the projection only
through what it needs. The two entry points the projection calls back into —
folding a call, folding a builtin — are the coupling that remains, and it is the
honest one: a fold *is* a frame run that is not allowed to write.
